Upay (wip)

+++ app/Takaden/Orderable.php
@@ -8,7 +8,7 @@ interface Orderable
 {
     public function handleSuccessPayment(Payable $payment);
     public function handleFailPayment(Payable $payment);
-    public function getTakadenAmount(): int;
+    public function getTakadenAmount(): float;
     public function getTakadenCurrency(): string;
     public function getTakadenUniqueId(): string;
     public function getTakadenRedirectUrl(): string;
